chore: migrate benchmark script to TypeScript
